Handle Chris bug #77:
when I follow the cpanel instructions after brand new installation: It does not tell me to give all permissions to the user as I add it to the database. I looked at the linux instructions and found it, but for completeness, might want to add that step to the instructions.